requirements.txt
altgraph==0.17.4
Babel==2.15.0
bcrypt==4.1.3
chardet==5.2.0
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
et-xmlfile==1.1.0
openpyxl==3.1.2
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pypdf==4.2.0
pypng==0.20220715.0
pywin32-ctypes==0.2.2
qrcode==7.4.2
reportlab==4.2.0
setuptools==70.0.0
tkcalendar==1.6.1
typing_extensions==4.12.0
workbook==1.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
